improve dependency definitions on produced crates
When uploaded to crates.io, crates lose any path dependency in
Cargo.toml. In the case, version information should be added in advance
so that crates.io pick it up.

This commit is a preparation to upload all of included crates to
crate.io. All of dependencies to included crate have both of path
dependency and version information planned on crates.io.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>